Integrated zcmplay into makezcm, rename to zcmtool

+++ tools/zcmtool.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Convert source audio file into ZCM digi file.
+# Tool to convert source audio file into ZCM digi file, or play them, or show info.
 # Requires python 'sox' module to be installed (pip install sox), and the sox command line tool.
 
 import os
@@ -11,7 +11,6 @@ import argparse
 
 
 class ZcmHeader():
-    # note: this class is shared with other Zcm tools
     MAX_SIZE = 2048 * 1024 - 8192  # 2 mb banked ram minus 1 bank (8kb) for kernal.
     HEADER_FORMAT = "<HBHBBB"
 
@@ -103,22 +102,66 @@ def convert_raw_to_zcm(raw_file: str, output_file: str, outputformat: dict) -> N
         print("Note: actual sample rate differs from requested sample rate (due to Vera clock).")
 
 
+def play_audio_file(filename: str, only_info: bool = True) -> None:
+    if os.path.splitext(filename)[1].lower() == ".zcm":
+        # play ZCM file by reading the raw pcm data
+        with open(filename, "rb") as fin:
+            header = ZcmHeader.from_header(fin.read(8))
+            pcm_bytes = fin.read()
+            assert header.size == len(pcm_bytes)
+        print(f"Info about zcm file {filename}")
+        print(f"pcm size: {header.size} bytes")
+        print(f"bits: {header.bits}")
+        print(f"channels: {header.channels}")
+        print(f"sample rate: {header.sample_rate}")
+        print(f"volume: {header.volume}")
+        print(f"duration: {round(header.duration, 3)} sec.")
+        if not only_info:
+            print("Playing audio.")
+            with tempfile.NamedTemporaryFile(suffix=".raw") as tmpfile:
+                tmpfile.write(pcm_bytes)
+                sox.core.sox(["-t", "raw", "-r", str(header.sample_rate),
+                              "-b", str(header.bits),
+                              "-c", str(header.channels),
+                              "-e", "signed-integer", tmpfile.name, "-d"])
+    else:
+        # let sox handle the playback of the audio file
+        info = sox.file_info.info(filename)
+        print(f"Info about audio file {filename}")
+        print(f"encoding: {info['encoding']}")
+        print(f"bits: {info['bitdepth']}")
+        print(f"channels: {info['channels']}")
+        print(f"sample rate: {info['sample_rate']}")
+        print(f"duration: {info['duration']} sec.")
+        if not only_info:
+            print("Playing audio.")
+            sox.core.play([filename])
+
+
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="convert audio to zcm digi")
+    parser = argparse.ArgumentParser(description="zcm digi audio conversion and playback")
     parser.add_argument("-s", "--stereo", action="store_true", help="stereo output, default is mono")
     parser.add_argument("-r", "--rate", type=int, help="output sample rate, default is %(default)s", default=22050)
     parser.add_argument("-8", "--eight", action="store_true", help="8 bits output, default is 16")
     parser.add_argument("-o", "--output", type=str, help="provide your own output file name")
-    parser.add_argument("--raw", action="store_true", help="input file is raw pcm already, just add zcm header")
+    parser.add_argument("-p", "--play", action="store_true", help="playback the input file")
+    parser.add_argument("-i", "--info", action="store_true", help="show info about the input file")
+    parser.add_argument("--raw", action="store_true", help="input file is raw pcm, just add zcm header when converting")
     parser.add_argument("inputfile")
     args = parser.parse_args()
-    outputformat = {
-        "channels": 2 if args.stereo else 1,
-        "rate": args.rate,
-        "bits": 8 if args.eight else 16
-    }
-    output_file = args.output if args.output else (os.path.splitext(os.path.split(args.inputfile)[1])[0] + ".zcm").upper()
-    raw_file = args.inputfile if args.raw else convert_source_to_raw(args.inputfile, outputformat)
-    convert_raw_to_zcm(raw_file, output_file, outputformat)
-    if not args.raw:
-        os.remove(raw_file)
+    if args.play:
+        play_audio_file(args.inputfile, only_info=False)
+    elif args.info:
+        play_audio_file(args.inputfile, only_info=True)
+    else:
+        # audio conversion
+        outputformat = {
+            "channels": 2 if args.stereo else 1,
+            "rate": args.rate,
+            "bits": 8 if args.eight else 16
+        }
+        output_file = args.output if args.output else (os.path.splitext(os.path.split(args.inputfile)[1])[0] + ".zcm").upper()
+        raw_file = args.inputfile if args.raw else convert_source_to_raw(args.inputfile, outputformat)
+        convert_raw_to_zcm(raw_file, output_file, outputformat)
+        if not args.raw:
+            os.remove(raw_file)
